==> scripts/m@hr/2_EFA.Rmd <==
---
title: "2_EFA"
author: "Nick Kathios"
date: '2022-10-31'
output: html_document
---
#load packages
```{r}
library(tidyverse)
library(psych)
library(corrplot)
library(moments)
source("../reverseCode.R")
library(psych)
library(EFAtools)
library(EFA.dimensions)
library(corrplot)
```

#read in data
```{r}
data<-read.csv(file='../../data/pleasing-displeasing_followup/scored_data.csv') 
original <- read.csv(file='../../data/pleasing-displeasing_followup/pleasing_displeasing_scored.csv')
MH <- read.csv(file='../../data/pleasing-displeasing_followup/scored_data.csv') %>%
  dplyr::select("M.H1_1":"M.H35_1")

#removing check questions
MH <- MH[-c(32:34)]

cols_reverse <- c("M.H6_1","M.H7_1","M.H12_1","M.H14_1","M.H17_1","M.H18_1","M.H20_1","M.H23_1","M.H29_1","M.H30_1",
                  "M.H32_1")
MH[cols_reverse] <- lapply(MH[cols_reverse], reverseCode, min=1, max=7)

updateMH <- MH[complete.cases(MH),]

names(original)[2] <- "ID"
demo <- merge(original, data, by.x = "ID", by.y="Q158")

demo1 <- demo[complete.cases(demo[500:530]),]

table(demo1$What.is.your.ethnicity..check.all.that.apply..)
```

#determine how many factors to use
```{r}
#initial determination of factor structure
#Kaiser's criterion
fafitfree <- fa(updateMH,nfactors = 6, rotate = "oblimin",covar = FALSE, SMC = TRUE,
 fm="ml", 
 max.iter=100)
n_factors <- length(fafitfree$e.values)
scree     <- data.frame(
  Factor_n =  as.factor(1:n_factors), 
  Eigenvalue = fafitfree$e.values)

ggplot(scree, aes(x = Factor_n, y = Eigenvalue, group = 1)) + 
  geom_point() + geom_line() +
  xlab("Number of factors") +
  ylab("Initial eigenvalue") +
  labs( title = "Scree Plot", 
        subtitle = "(Based on the unreduced correlation matrix)")+
  geom_line(y=1,color="red") #suggests 6

#parallel analysis
parallel <- fa.parallel(updateMH)

#MAP
MAP <- vss(updateMH) #suggests 6 factors
MAP <- MAP[["map"]]
print(MAP)

#McDonald Omega Test on 6 factors
fa.none <- psych::fa(r=updateMH, 
 nfactors = 6, 
  covar = FALSE, SMC = TRUE,
 fm="ml", 
 max.iter=1000,  
 rotate="oblimin") # no rotation

print(fa.none)

fa.diagram(fa.none)

SL_transformation <- EFAtools::SL(fa.none, type="psych",method="ML")
SL_transformation <- SL_transformation[[1]]

omega <-omega(SL_transformation, nfactors=6,n.iter=1,p=.05,poly=FALSE,key=NULL,
    flip=TRUE,digits=2, title="Omega",sl=TRUE,labels=NULL,
    plot=TRUE)

summary(omega)

#extracting factors
#FA w/ one factor
fa.none <- psych::fa(r=updateMH, 
 nfactors = 1, 
  covar = FALSE, SMC = TRUE,
 fm="ml", 
 max.iter=100,  
 rotate="varimax")

#parallel analysis on residual to find factor #
residual <- fa.none[["residual"]]
parallel <- fa.parallel(residual) #suggests 4 factors
MAP <- vss(residual) #suggests 5 factors
MAP <- MAP[["map"]]
print(MAP)

fafitfree <- fa(residual,nfactors = ncol(residual), rotate = "oblimin",covar = FALSE, SMC = TRUE,
 fm="ml", 
 max.iter=100)
n_factors <- length(fafitfree$e.values)
scree     <- data.frame(
  Factor_n =  as.factor(1:n_factors), 
  Eigenvalue = fafitfree$e.values)

ggplot(scree, aes(x = Factor_n, y = Eigenvalue, group = 1)) + 
  geom_point() + geom_line() +
  xlab("Number of factors") +
  ylab("Initial eigenvalue") +
  labs( title = "Scree Plot", 
        subtitle = "(Based on the unreduced correlation matrix)")+
  geom_line(y=1,color="red") #suggests 7

#item reduction on this factor structure
fa.none <- psych::fa(r=updateMH, 
 nfactors = 7, 
  covar = FALSE, SMC = TRUE,
 fm="ml", 
 max.iter=100,  
 rotate="oblimin") # no rotation
summary(fa.none)
SL_transformation <- EFAtools::SL(fa.none, type="psych",method="ML")
SL_transformation <- SL_transformation[[2]]
SL_transformation

#remove items
updateMH <- dplyr::select(updateMH, -c("M.H17_1", "M.H20_1", "M.H18_1","M.H19_1","M.H13_1","M.H27_1","M.H23_1"))

#determine new no. of factors
#FA w/ one factor
fa.none <- psych::fa(r=updateMH, 
 nfactors = 1, 
  covar = FALSE, SMC = TRUE,
 fm="ml", 
 max.iter=100,  
 rotate="varimax")

#parallel analysis on residual to find factor #
residual <- fa.none[["residual"]]
parallel <- fa.parallel(residual) #suggests 4
MAP <- vss(residual)
MAP <- MAP[["map"]]
print(MAP) #suggests 4

fafitfree <- fa(residual,nfactors = ncol(residual), rotate = "oblimin",covar = FALSE, SMC = TRUE,
 fm="ml", 
 max.iter=100)
n_factors <- length(fafitfree$e.values)
scree     <- data.frame(
  Factor_n =  as.factor(1:n_factors), 
  Eigenvalue = fafitfree$e.values)

ggplot(scree, aes(x = Factor_n, y = Eigenvalue, group = 1)) + 
  geom_point() + geom_line() +
  xlab("Number of factors") +
  ylab("Initial eigenvalue") +
  labs( title = "Scree Plot", 
        subtitle = "(Based on the unreduced correlation matrix)")+
  geom_line(y=1,color="red") #suggests 6

#item reduction on this factor structure (6 has best BIC + RSMEA)
fa.none <- psych::fa(r=updateMH, 
 nfactors = 6, 
  covar = FALSE, SMC = TRUE,
 fm="ml", 
 max.iter=100,  
 rotate="oblimin") # no rotation
summary(fa.none)
SL_transformation <- EFAtools::SL(fa.none, type="psych",method="ML")
SL_transformation <- SL_transformation[[2]]
SL_transformation

#run again
updateMH <- dplyr::select(updateMH, -c("M.H15_1", "M.H24_1"))

#determine new no. of factors
#FA w/ one factor
fa.none <- psych::fa(r=updateMH, 
 nfactors = 1, 
  covar = FALSE, SMC = TRUE,
 fm="ml", 
 max.iter=100,  
 rotate="varimax")

#parallel analysis on residual to find factor #
residual <- fa.none[["residual"]]
parallel <- fa.parallel(residual) #suggests 4
MAP <- vss(residual)
MAP <- MAP[["map"]] #suggests 4
print(MAP)

fafitfree <- fa(residual,nfactors = ncol(residual), rotate = "oblimin",covar = FALSE, SMC = TRUE,
 fm="ml", 
 max.iter=100)
n_factors <- length(fafitfree$e.values)
scree     <- data.frame(
  Factor_n =  as.factor(1:n_factors), 
  Eigenvalue = fafitfree$e.values)

ggplot(scree, aes(x = Factor_n, y = Eigenvalue, group = 1)) + 
  geom_point() + geom_line() +
  xlab("Number of factors") +
  ylab("Initial eigenvalue") +
  labs( title = "Scree Plot", 
        subtitle = "(Based on the unreduced correlation matrix)")+
  geom_line(y=1,color="red") #suggests 5 factors

#item reduction on this factor structure
fa.none <- psych::fa(r=updateMH, 
 nfactors = 5, 
  covar = FALSE, SMC = TRUE,
 fm="ml", 
 max.iter=100,  
 rotate="oblimin") # no rotation
summary(fa.none)
SL_transformation <- EFAtools::SL(fa.none, type="psych",method="ML")
SL_transformation <- SL_transformation[[2]]
SL_transformation

#run again
updateMH <- dplyr::select(updateMH, -c("M.H22_1", "M.H31_1"))

#determine new no. of factors
#FA w/ one factor
fa.none <- psych::fa(r=updateMH, 
 nfactors = 1, 
  covar = FALSE, SMC = TRUE,
 fm="ml", 
 max.iter=100,  
 rotate="varimax")

#parallel analysis on residual to find factor #
residual <- fa.none[["residual"]]
parallel <- fa.parallel(residual) #suggests 3
MAP <- vss(residual)
MAP <- MAP[["map"]]
print(MAP) #suggests 4

fafitfree <- fa(residual,nfactors = ncol(residual), rotate = "oblimin",covar = FALSE, SMC = TRUE,
 fm="ml", 
 max.iter=100)
n_factors <- length(fafitfree$e.values)
scree     <- data.frame(
  Factor_n =  as.factor(1:n_factors), 
  Eigenvalue = fafitfree$e.values)

ggplot(scree, aes(x = Factor_n, y = Eigenvalue, group = 1)) + 
  geom_point() + geom_line() +
  xlab("Number of factors") +
  ylab("Initial eigenvalue") +
  labs( title = "Scree Plot", 
        subtitle = "(Based on the unreduced correlation matrix)")+
  geom_line(y=1,color="red") #suggests 5

#item reduction on this factor structure
fa.none <- psych::fa(r=updateMH, 
 nfactors = 5, 
  covar = FALSE, SMC = TRUE,
 fm="ml", 
 max.iter=100,  
 rotate="oblimin") # no rotation
summary(fa.none)
SL_transformation <- EFAtools::SL(fa.none, type="psych",method="ML")
SL_transformation <- SL_transformation[[2]]
SL_transformation

omega <-omega(SL_transformation, nfactors=5,n.iter=1,p=.05,poly=FALSE,key=NULL,
    flip=TRUE,digits=2, title="Omega",sl=TRUE,labels=NULL,
    plot=TRUE)

summary(omega)
```

#scoring + combining data (maybe make into a new script / only run once - ALREADY RAN)
```{r}
scored_data<-read.csv(file='../../data/pleasing-displeasing_followup/scored_data.csv') 
master <- read.csv(file="../../data/pleasing-displeasing_followup/pleasing_displeasing_scored.csv")

#names(scored_data)[20] <- "ID"
master <- master[-1]
names(master)[1] <- "ID"

#scoring
MH <- scored_data %>%
  dplyr::select("M.H1_1":"M.H35_1")
  
MH <- MH[-c(32:34)] %>%
  dplyr::select(-c("M.H17_1", "M.H20_1", "M.H18_1","M.H19_1","M.H13_1","M.H27_1","M.H23_1")) %>%
  dplyr::select( -c("M.H15_1", "M.H24_1")) %>%
  dplyr::select( -c("M.H22_1", "M.H31_1"))

#scoring
cols_reverse <- c("M.H6_1","M.H7_1","M.H12_1","M.H14_1","M.H29_1","M.H30_1",
                  "M.H32_1")
cols_caregiverbeliefs <- c("M.H1_1","M.H2_1","M.H3_1","M.H4_1")
cols_caregiversinging <- c("M.H9_1","M.H10_1","M.H12_1","M.H16_1")
cols_childhoodattitude <- c("M.H29_1","M.H30_1","M.H32_1")
cols_socialcontexts <- c("M.H21_1","M.H25_1","M.H26_1","M.H28_1")
cols_musicmaking <- c("M.H5_1","M.H6_1","M.H7_1","M.H8_1", "M.H14_1")

# Reverse code
MH[cols_reverse] <- lapply(MH[cols_reverse], reverseCode,  min=1, max=7)

#scoring
scored_data$MH_total <- apply(MH,1, sum) # total BMRQ scores
scored_data$MH_caregiverbeliefs <- apply(MH[cols_caregiverbeliefs], 1,sum)
scored_data$MH_caregiversinging <- apply(MH[cols_caregiversinging],1, sum)
scored_data$MH_childhoodattitude <-apply(MH[cols_childhoodattitude],1, sum)
scored_data$MH_socialcontexts <- apply(MH[cols_socialcontexts],1, sum)
scored_data$MH_musicmaking <- apply(MH[cols_musicmaking],1, sum)


full <- merge(scored_data, master,by.x="Q158", by.y="ID")

full <- full[complete.cases(full[199:228]),]

write.csv(scored_data, file='../../data/pleasing-displeasing_followup/EFA_scored_data.csv')

write.csv(full, file='../../data/pleasing-displeasing_followup/scored_data_mhcomplete.csv')
```
