==> scripts/m@hr/3_validity.Rmd <==
---
title: "3_validity"
author: "Nick Kathios"
date: '2023-01-17'
output: html_document
---

#load packages
```{r}
library(tidyverse)
library(corrplot)
library(psych)
source("../reverseCode.R")
# R command to get corrplot2 from github
source("https://raw.githubusercontent.com/dorianps/corrplot2/master/corrplot2.R")
```

#read in data
```{r}
full <- read.csv(file='../../data/pleasing-displeasing_followup/scored_data_mhcomplete.csv')
```

#corr matrices --> START HERE
```{r}
#M@HR + GoldMSI
variables = dplyr::select(full, 
                          MH_total, MH_caregiverbeliefs, MH_caregiversinging, MH_childhoodattitude, MH_socialcontexts,
                          MH_musicmaking, Gold_activeengagement, Gold_perceptualabilities, Gold_musicaltraining,
                          Gold_singingabilities, Gold_emotion, Gold_musicalsophistication,BMRQ_total, BMRQ_moodreg, BMRQ_emotion,
                          BMRQ_socialreward,BMRQ_absorption, BMRQ_musicseek, BMRQ_sensorimotor,QUIC_child_total, 
                          CHAOS_child_total, deprivation_child_total, threat_childhood_total)
output = corr.test(variables)

corr_matrix = output$r
p_matrix <- output$p

colnames(corr_matrix) <- c('1', '2', '3', '4', '5', '6', '7', "8", "9", "10", "11", "12",'13', '14', '15', '16', '17', '18',
                         '19',"20", "21", "22", "23")
rownames(corr_matrix) <- c('1', '2', '3', '4', '5', '6', '7', "8", "9", "10", "11", "12",'13', '14', '15', '16', '17', '18',
                         '19',"20", "21", "22", "23")


pdf(file = "../../figures/pleasing:displeasing_followup/EFA_corrplot_totalsnobold.pdf", width = 13, height=13)

## update to get corrplot2
corrplot(corr_matrix, p.mat = p_matrix, insig = "label_sig",
         sig.level = c(0.001,0.01,0.05), pch.cex=0.9, pch.col = "black", tl.col = "black", type="lower",tl.srt =45, 
         col=colorRampPalette(c("darkblue","white","red"))(100),cl.lim=c(0,1), diag = FALSE,tl.cex = 1, cl.cex = 1)
dev.off()


#M@HR + ELA measures
variables = dplyr::select(full, 
                          MH_total, MH_caregiverbeliefs, MH_caregiversinging, MH_childhoodattitude, MH_socialcontexts, MH_musicmaking, QUIC_child_total, 
                          CHAOS_child_total, deprivation_child_total, threat_childhood_total)
output = corr.test(variables)

corr_matrix = output$r
p_matrix <- output$p

colnames(corr_matrix) <- c('M@H-R General Score', 'Caregiver Beliefs', 'Caregiver Singing', 'Attitude to Childhood Music', 'Social Contexts', 'Music-Making Tendencies', "QUIC Childhood", "CHAOS Childhood", "Deprivation Childhood", "Threat Childhood")
rownames(corr_matrix) <- c('M@H-R General Score', 'Caregiver Beliefs', 'Caregiver Singing', 'Attitude to Childhood Music', 'Social Contexts', 'Music-Making Tendencies', "QUIC Childhood", "CHAOS Childhood", "Deprivation Childhood", "Threat Childhood")

pdf(file = "../../figures/pleasing:displeasing_followup/BMRQ_corrplot_totals.pdf")

corrplot(corr_matrix, p.mat = p_matrix, insig = "label_sig",
         sig.level = c(0.001,0.01,0.05), pch.cex=0.9, pch.col = "black", tl.col = "black", type="lower",tl.srt =45, 
         col=colorRampPalette(c("darkblue","white","red"))(100),cl.lim=c(0,1))

dev.off()
```

#modeling stuff
```{r}
m1 <- lm(data=full, BMRQ_socialreward ~ MH_caregiverbeliefs + MH_caregiversinging + MH_childhoodattitude + MH_socialcontexts + MH_musicmaking)
summary(m1)

m1 <- lm(data=full, BMRQ_total ~ MH_caregiverbeliefs + MH_caregiversinging + MH_childhoodattitude + MH_socialcontexts + MH_musicmaking)
summary(m1)

m1 <- lm(data=full, HUMS_adult_healthy ~ MH_caregiverbeliefs + MH_caregiversinging + MH_childhoodattitude + MH_socialcontexts + MH_musicmaking)
summary(m1)

m1 <- lm(data=full, HUMS_adult_unhealthy ~ MH_caregiverbeliefs + MH_caregiversinging + MH_childhoodattitude + MH_socialcontexts + MH_musicmaking)
summary(m1)

m1 <- lm(data=full, scale(MH_socialcontexts) ~ scale(BMRQ_moodreg) + scale(BMRQ_emotion) + scale(BMRQ_socialreward) + scale(BMRQ_absorption) + scale(BMRQ_musicseek) + scale(BMRQ_sensorimotor))
summary(m1)
```

#updated figures for manuscript 9/25
```{r}
#GoldMSI and M@H
variables = dplyr::select(full, 
                          MH_total, MH_caregiverbeliefs, MH_caregiversinging, MH_childhoodattitude, MH_socialcontexts,
                          MH_musicmaking, Gold_activeengagement, Gold_perceptualabilities, Gold_musicaltraining,
                          Gold_singingabilities, Gold_emotion, Gold_musicalsophistication)
output = corr.test(variables)

cor.test(full$MH_total,full$Gold_activeengagement)

corr_matrix = output$r
p_matrix <- output$p

colnames(corr_matrix) <- c('M@HR General', 'M@HR Caregiver Beliefs', 'M@HR Caregiver Singing', 'M@HR Attitude', 'M@HR Social Contexts', 'M@HR Child Engagement', 'Active Engagment', "Perceptual Abilities", "Musical Training", "Singing Abilities", "Emotion", "Musical Sophistication")
rownames(corr_matrix) <- c('M@HR General', 'M@HR Caregiver Beliefs', 'M@HR Caregiver Singing', 'M@HR Attitude', 'M@HR Social Contexts', 'M@HR Child Engagement', 'Active Engagment', "Perceptual Abilities", "Musical Training", "Singing Abilities", "Emotion", "Musical Sophistication")

pdf(file = "../../figures/m@h/Gold_Study1.pdf")

## update to get corrplot2
corrplot(corr_matrix, p.mat = p_matrix, insig = "label_sig",
         sig.level = c(0.001,0.01,0.05), pch.cex=0.9, pch.col = "black", tl.col = "black", type="lower",tl.srt =45, 
         col=colorRampPalette(c("darkblue","white","red"))(100),cl.lim=c(0,1), diag = FALSE,tl.cex = 1, cl.cex = 1)

dev.off()

#eBMRQ and M@H
variables = dplyr::select(full, 
                          MH_total, MH_caregiverbeliefs, MH_caregiversinging, MH_childhoodattitude, MH_socialcontexts,
                          MH_musicmaking, BMRQ_total, BMRQ_moodreg, BMRQ_emotion,
                          BMRQ_socialreward,BMRQ_absorption, BMRQ_musicseek, BMRQ_sensorimotor)
output = corr.test(variables)

corr_matrix = output$r
p_matrix <- output$p

colnames(corr_matrix) <- c('M@HR General', 'M@HR Caregiver Beliefs', 'M@HR Caregiver Singing', 'M@HR Attitude', 'M@HR Social Contexts', 'M@HR Child Engagement', 'BMRQ Total', "Mood Regulation", "Emotion", "Social Reward", "Absorption", "Music Seeking", "Sensorimotor")
rownames(corr_matrix) <- c('M@HR General', 'M@HR Caregiver Beliefs', 'M@HR Caregiver Singing', 'M@HR Attitude', 'M@HR Social Contexts', 'M@HR Child Engagement', 'BMRQ Total', "Mood Regulation", "Emotion", "Social Reward", "Absorption", "Music Seeking", "Sensorimotor")

pdf(file = "../../figures/m@h/eBMRQ_Study1.pdf")

## update to get corrplot2
corrplot(corr_matrix, p.mat = p_matrix, insig = "label_sig",
         sig.level = c(0.001,0.01,0.05), pch.cex=0.9, pch.col = "black", tl.col = "black", type="lower",tl.srt =45, 
         col=colorRampPalette(c("darkblue","white","red"))(100),cl.lim=c(0,1), diag = FALSE,tl.cex = 1, cl.cex = 1)

dev.off()
```

#Correlation values
```{r}
mh <- full$MH_childhoodattitude
adv <- full$deprivation_child_total

cor.test(mh, adv)
```
