==> scripts/m@hr/4_CFA.Rmd <==
---
title: "4_CFA"
author: "Nick Kathios"
date: '2023-03-22'
output: html_document
---

```{r}
library(tidyverse)
source("../reverseCode.R")
library(psych)
library(corrplot)
library(lavaan)
library(semPlot)
library(sjPlot)
library(dplyr)
library(rempsyc)
```

#read in data
```{r}
first <- read.csv(file='../../data/pleasing-displeasing_followup/scored_data.csv') 
all <- read.csv(file='../../data/SLfollowupscored.csv')

#for tables
goldmsi_table <- read.csv(file='../../figures/m@h/goldmsi_table.csv')
ebmrq_table <- read.csv(file='../../figures/m@h/ebmrq_table.csv')
testretest_table <- read.csv(file='../../figures/m@h/testretest_table.csv')
SRQ_table <- read.csv(file='../../figures/m@h/srq_table.csv')
demo_table1 <- read.csv(file='../../data/m@hr/demo_table_study1.csv')
demo_table <- read.csv(file='../../data/m@hr/demo_table_study2.csv')
```

#correcting musical training scoring of goldmsi
```{r}
gold <- all[608:647]

#for those who reported no musical training, turning NAs to 1s
gold[is.na(gold)] <- 1

#for those who reported "no daily practice" (8th option to question), giving them a 1 score
gold$On.average..at.the.peak.of.my.interest..I.practiced.my.primary.instrument.this.many.hours.per.day....Selected.Choice <-
  ifelse(gold$On.average..at.the.peak.of.my.interest..I.practiced.my.primary.instrument.this.many.hours.per.day....Selected.Choice==8,1,gold$On.average..at.the.peak.of.my.interest..I.practiced.my.primary.instrument.this.many.hours.per.day....Selected.Choice)


cols_reverse <- c("Please.choose.the.most.appropriate.category....I.don.t.spend.much.of.my.disposable.income.on.music." , 
"Please.choose.the.most.appropriate.category....I.find.it.difficult.to.spot.mistakes.in.a.performance.of.a.song.even.if.I.know.the.tune." , 
"Please.choose.the.most.appropriate.category....I.have.trouble.recognizing.a.familiar.song.when.played.in.a.different.way.or.by.a.different.performer." , 
"Please.choose.the.most.appropriate.category....When.I.sing..I.have.no.idea.whether.I.m.in.tune.or.not." , 
"Please.choose.the.most.appropriate.category....I.have.never.been.complimented.for.my.talents.as.a.musical.performer." , 
"Please.choose.the.most.appropriate.category....I.would.not.consider.myself.a.musician." , 
"Please.choose.the.most.appropriate.category....I.am.not.able.to.sing.in.harmony.when.somebody.is.singing.a.familiar.tune." , 
"Please.choose.the.most.appropriate.category....I.don.t.like.singing.in.public.because.I.m.afraid.that.I.would.sing.wrong.notes." , 
"Please.choose.the.most.appropriate.category....Pieces.of.music.rarely.evoke.emotions.for.me.")

cols_musicalsophistication <- c("Please.choose.the.most.appropriate.category....I.spend.a.lot.of.free.time.doing.music.related.activities." , 
           "Please.choose.the.most.appropriate.category....I.enjoy.writing.about.music..for.example.on.blogs.and.forums." , 
"Please.choose.the.most.appropriate.category....If.somebody.starts.singing.a.song.I.don.t.know..I.can.usually.join.in." , 
"Please.choose.the.most.appropriate.category....I.can.sing.or.play.music.from.memory." , 
"Please.choose.the.most.appropriate.category....I.am.able.to.hit.the.right.notes.when.I.sing.along.with.a.recording." , 
"Please.choose.the.most.appropriate.category....I.can.compare.and.discuss.differences.between.two.performances.or.versions.of.the.same.piece.of.music." , 
"Please.choose.the.most.appropriate.category....I.have.never.been.complimented.for.my.talents.as.a.musical.performer." , 
"Please.choose.the.most.appropriate.category....I.often.read.or.search.the.internet.for.things.related.to.music." , 
"Please.choose.the.most.appropriate.category....I.am.not.able.to.sing.in.harmony.when.somebody.is.singing.a.familiar.tune." , 
"Please.choose.the.most.appropriate.category....I.am.able.to.identify.what.is.special.about.a.given.musical.piece." , 
"Please.choose.the.most.appropriate.category....When.I.sing..I.have.no.idea.whether.I.m.in.tune.or.not." , 
"Please.choose.the.most.appropriate.category....Music.is.kind.of.an.addiction.for.me..I.couldn.t.live.without.it." , 
"Please.choose.the.most.appropriate.category....I.don.t.like.singing.in.public.because.I.m.afraid.that.I.would.sing.wrong.notes." , 
"Please.choose.the.most.appropriate.category....I.would.not.consider.myself.a.musician." , 
"Please.choose.the.most.appropriate.category....After.hearing.a.new.song.two.or.three.times..I.can.usually.sing.it.by.myself." , 
"I.engaged.in.regular..daily.practice.of.a.musical.instrument..including.voice..for." , 
"On.average..at.the.peak.of.my.interest..I.practiced.my.primary.instrument.this.many.hours.per.day....Selected.Choice" , "I.can.play.this.many.musical.instruments.")

cols_musicaltraining <- c("Please.choose.the.most.appropriate.category....I.have.never.been.complimented.for.my.talents.as.a.musical.performer." , "Please.choose.the.most.appropriate.category....I.would.not.consider.myself.a.musician.", 
"I.engaged.in.regular..daily.practice.of.a.musical.instrument..including.voice..for." , 
"On.average..at.the.peak.of.my.interest..I.practiced.my.primary.instrument.this.many.hours.per.day....Selected.Choice" , 
"I.have.had.formal.training.in.music.theory.for." , 
"During.my.lifetime..I.have.had.formal.training.on.a.music.instrument..including.voice..for." , 
"I.can.play.this.many.musical.instruments.")


# Reverse code
gold[cols_reverse] <- lapply(gold[cols_reverse], reverseCode, min=1, max=7)

# Add sub-scores and total to main df

all$Gold_musicaltraining <-apply(gold[cols_musicaltraining],1, sum)/7
all$Gold_musicalsophistication <- apply(gold[cols_musicalsophistication],1, sum)/18
```


```{r}
keep <- setdiff(all$ID,first$Q158)

second <- all[all$ID %in% keep, ] 

MH_second <- second %>%
  dplyr::select("M.H_child1_1":"M.H_child20_1")

MH_second <- MH_second[complete.cases(MH_second),]

cols_reverse <- c("M.H_child6_1","M.H_child7_1","M.H_child11_1","M.H_child12_1","M.H_child18_1","M.H_child19_1","M.H_child20_1")
MH_second[cols_reverse] <- lapply(MH_second[cols_reverse], reverseCode, min=1, max=7)

MH_all <- all %>%
  dplyr::select("M.H_child1_1":"M.H_child20_1")

MH_all <- MH_all[complete.cases(MH_all),]

cols_reverse <- c("M.H_child6_1","M.H_child7_1","M.H_child11_1","M.H_child12_1","M.H_child18_1","M.H_child19_1","M.H_child20_1")
MH_all[cols_reverse] <- lapply(MH_all[cols_reverse], reverseCode, min=1, max=7)

#use "all" for *new* construct validity results (SRQ + MET)
#use "second" for replications of construct validity results from the EFA, as well as the CFA
#the retest df is created below

range(second$What.is.your.age.)
```

#CFA
```{r}
model <- 'G =~ M.H_child10_1 + M.H_child1_1 + M.H_child2_1 + M.H_child3_1 + M.H_child4_1 + M.H_child9_1  + M.H_child11_1  + M.H_child13_1 + M.H_child18_1 + M.H_child19_1 + M.H_child20_1 + M.H_child14_1 + M.H_child15_1 + M.H_child16_1 + M.H_child17_1 + M.H_child5_1 + M.H_child6_1 + M.H_child7_1 + M.H_child8_1 + M.H_child12_1
  
  caregiverbeliefs   =~ M.H_child2_1 + M.H_child1_1 +  + M.H_child3_1 + M.H_child4_1 
  caregiversinging =~ M.H_child10_1 + M.H_child9_1  + M.H_child11_1 + M.H_child13_1 
  childhoodattitude   =~ M.H_child20_1 + M.H_child19_1 + M.H_child18_1  
  socialcontexts  =~  M.H_child15_1 + M.H_child14_1 + M.H_child16_1 + M.H_child17_1
  musicmaking =~ M.H_child7_1 + M.H_child5_1 + M.H_child6_1  + M.H_child8_1 + M.H_child12_1

## restrictions --> make sure factors are uncorrelated
G ~~ 0*caregiverbeliefs
G ~~ 0*caregiversinging
G ~~ 0*childhoodattitude
G ~~ 0*socialcontexts
G ~~ 0*musicmaking

caregiverbeliefs ~~ 0*caregiversinging
caregiverbeliefs ~~ 0*childhoodattitude
caregiverbeliefs ~~ 0*socialcontexts
caregiverbeliefs ~~ 0*musicmaking

caregiversinging ~~ 0*childhoodattitude
caregiversinging ~~ 0*socialcontexts
caregiversinging ~~ 0*musicmaking

childhoodattitude ~~ 0*socialcontexts
childhoodattitude ~~ 0*musicmaking

socialcontexts ~~ 0*musicmaking

#residual variances --> setting to 1
    G ~~ 1*G
    caregiverbeliefs ~~ 1*caregiverbeliefs
    caregiversinging ~~ 1*caregiversinging
    childhoodattitude ~~ 1*childhoodattitude
    socialcontexts ~~ 1*socialcontexts
    musicmaking ~~ 1*musicmaking'

fitG <- lavaan::cfa(model, data = MH_all,std.lv=TRUE)
lavaan::summary(fitG, fit.measures = T, standardized = T)
semPaths(fitG)
```

```{r}
#scoring
cols_caregiverbeliefs <- c("M.H_child1_1","M.H_child2_1","M.H_child3_1","M.H_child4_1")
cols_caregiversinging <- c("M.H_child9_1","M.H_child10_1","M.H_child11_1","M.H_child13_1")
cols_childhoodattitude <- c("M.H_child18_1","M.H_child19_1","M.H_child20_1")
cols_socialcontexts <- c("M.H_child14_1","M.H_child15_1","M.H_child16_1","M.H_child17_1")
cols_musicmaking <- c("M.H_child5_1","M.H_child6_1","M.H_child7_1","M.H_child8_1", "M.H_child12_1")


#scoring
second$MH_total <- apply(MH_second,1, sum) # total BMRQ scores
second$MH_caregiverbeliefs <- apply(MH_second[cols_caregiverbeliefs], 1,sum)
second$MH_caregiversinging <- apply(MH_second[cols_caregiversinging],1, sum)
second$MH_childhoodattitude <-apply(MH_second[cols_childhoodattitude],1, sum)
second$MH_socialcontexts <- apply(MH_second[cols_socialcontexts],1, sum)
second$MH_musicmaking <- apply(MH_second[cols_musicmaking],1, sum)

all$MH_total <- apply(MH_all,1, sum) # total BMRQ scores
all$MH_caregiverbeliefs <- apply(MH_all[cols_caregiverbeliefs], 1,sum)
all$MH_caregiversinging <- apply(MH_all[cols_caregiversinging],1, sum)
all$MH_childhoodattitude <-apply(MH_all[cols_childhoodattitude],1, sum)
all$MH_socialcontexts <- apply(MH_all[cols_socialcontexts],1, sum)
all$MH_musicmaking <- apply(MH_all[cols_musicmaking],1, sum)


#validity tests
variables = dplyr::select(second, 
                          MH_total,MH_caregiverbeliefs,
                          MH_caregiversinging, MH_childhoodattitude, MH_musicmaking, MH_socialcontexts,
                          MET_melodicscore, MET_rhythmicscore, SRQ_adol_total,
                          SRQ_current_total)
output = corr.test(variables)

corr_matrix = output$r
p_matrix <- output$p

colnames(corr_matrix) <- c("1","2", "3",
                           "4", "5", 
                           "6","7", "8", 
                           "9", "10")
rownames(corr_matrix) <- c("1","2", "3",
                           "4", "5", 
                           "6","7", "8", 
                           "9", "10")

pdf(file = "../../figures/statistical_learningfollowup/CFA_new.pdf",width = 13, height=13)

corrplot(corr_matrix, p.mat = p_matrix, insig = "label_sig",
         sig.level = c(0.001,0.01,0.05), pch.cex=0.9, pch.col = "black", tl.col = "black", type="lower",tl.srt =45, 
         col=colorRampPalette(c("darkblue","white","red"))(100),cl.lim=c(0,1), diag = FALSE,tl.cex = 3, cl.cex = 2)

dev.off()

m1 <- lm(CD.RISC_total ~ MH_caregiverbeliefs + MH_caregiversinging + MH_childhoodattitude + MH_musicmaking + MH_socialcontexts, data=second)
summary(m1)


#how are the subscales relating to one another?
variables = dplyr::select(second, 
                          MH_total,MH_caregiverbeliefs,
                          MH_caregiversinging, MH_childhoodattitude, MH_musicmaking, MH_socialcontexts)
output = corr.test(variables)

corr_matrix = output$r
p_matrix <- output$p

pdf(file = "../../figures/statistical_learningfollowup/CFA_new.pdf",width = 13, height=13)

corrplot(corr_matrix, p.mat = p_matrix, insig = "label_sig",
         sig.level = c(0.001,0.01,0.05), pch.col = "black", tl.col = "black", type="lower",tl.srt =45, 
         col=colorRampPalette(c("darkblue","white","red"))(100),cl.lim=c(0,1), diag = FALSE)

dev.off()

#grabbing specific correlations across subscales
cor.test(all$MH_childhoodattitude, all$MH_caregiverbeliefs)
cor.test(all$MH_caregiverbeliefs, all$MH_caregiversinging)
cor.test(all$MH_caregiversinging, all$MH_musicmaking)
cor.test(all$MH_musicmaking, all$MH_socialcontexts)
cor.test(all$MH_socialcontexts, all$MH_total)
```


#test-retest correlations
```{r}
'%notin%' <- Negate('%in%')
retest <- subset(all, all$ID %notin% second$ID)
retest <- merge(retest, first, by.x = "ID", by.y = "PROLIFIC_PID") %>%
  dplyr::select(MH_childhoodattitude.x,MH_childhoodattitude.y)

ICC(retest, missing=TRUE, alpha=0.5, lmer=TRUE, check.keys = FALSE)

range(retest$What.is.your.age., na.rm = TRUE)
mean(retest$What.is.your.age., na.rm = TRUE)

table(retest$What.is.your.ethnicity..check.all.that.apply..)

#ICMPC Poster
MH_musicmaking.x <- retest$MH_musicmaking.x
MH_musicmaking.y <- retest$MH_musicmaking.y
corr.test(MH_musicmaking.x, MH_musicmaking.y)
```


```{r}
variables = dplyr::select(second, 
                          MH_total,
                          MH_caregiverbeliefs, MH_caregiversinging,
                          MH_childhoodattitude, MH_musicmaking, 
                          MH_socialcontexts, QUIC_child_total,
                          QUIC_child_parentalmonitoring, QUIC_child_parentalenvironment,
                          QUIC_child_parentalpredictability, QUIC_child_physiscalenvironment,
                          QUIC_child_safetyandsecurity, deprivation_child_total,
                          deprivation_child_emotional, deprivation_child_physical,
                          deprivation_child_cognitive, threat_childhood_total,
                          threat_childhood_home, threat_childhood_neighborhood,
                          CHAOS_child_total, CHAOS_child_shortform,
                          STAI_total, PASS_total, CD.RISC_total, HUMS_adult_healthy, HUMS_adult_unhealthy,
                          What.is.your.age., Gold_activeengagement,Gold_perceptualabilities, Gold_musicaltraining,
                          Gold_singingabilities, Gold_emotion, Gold_musicalsophistication,BMRQ_total, BMRQ_emotion, )
output = corr.test(variables)
corr_matrix <- output$r
p_matrix <- output$p
colnames(corr_matrix) <- c('General Score', 'Caregiver Beliefs', 'Caregiver Singing', 'Childhood Attitude', 'Childhood Engagement with Music', 'Social Contexts', 'QUIC Total', 'QUIC Parental Monitoring', 'QUIC Parental Environment', 'QUIC Parental Predictability', 'QUIC Physical Environment', 'QUIC Safety and Security', 'Deprivation Total', 'Deprivation Emotional', 'Deprivation Physical', 'Deprivation Cognitive', 'Threat Total', 'Threat Home', 'Threat Neighborhood', 'CHAOS Total', 'CHAOS Short Form', 'STAI', 'PASS', 'CD RISC', "Healthy Music Usage", "Unhealthy Music Usage", "Age", "GoldMSI - Active Engagement", "GoldMSI - Perceptual Abilities", "GoldMSI - Musical Training", "GoldMSI - Singing Abilities", "GoldMSI - Emotion", "GoldMSI - Musical Sophistication")
rownames(corr_matrix) <- c('General Score', 'Caregiver Beliefs', 'Caregiver Singing', 'Childhood Attitude', 'Childhood Engagement with Music', 'Social Contexts', 'QUIC Total', 'QUIC Parental Monitoring', 'QUIC Parental Environment', 'QUIC Parental Predictability', 'QUIC Physical Environment', 'QUIC Safety and Security', 'Deprivation Total', 'Deprivation Emotional', 'Deprivation Physical', 'Deprivation Cognitive', 'Threat Total', 'Threat Home', 'Threat Neighborhood', 'CHAOS Total', 'CHAOS Short Form', 'STAI', 'PASS','CD RISC', "Healthy Music Usage", "Unhealthy Music Usage", "Age","GoldMSI - Active Engagement", "GoldMSI - Perceptual Abilities", "GoldMSI - Musical Training", "GoldMSI - Singing Abilities", "GoldMSI - Emotion", "GoldMSI - Musical Sophistication")
corrplot(corr_matrix, p.mat = p_matrix, insig = "label_sig",
         sig.level = c(0.001,0.01,0.05), pch.cex=0.9, pch.col = "black", tl.col = "black", type="lower",tl.srt =45, 
         col=colorRampPalette(c("darkblue","white","red"))(100),cl.lim=c(0,1))
```

#For Figures 3 + 4 of manuscript
```{r}
#Replications
variables = dplyr::select(second, 
                          MH_total, MH_caregiverbeliefs, MH_caregiversinging, MH_childhoodattitude, MH_socialcontexts,
                          MH_musicmaking, Gold_activeengagement, Gold_perceptualabilities, Gold_musicaltraining,
                          Gold_singingabilities, Gold_emotion, Gold_musicalsophistication, BMRQ_total, BMRQ_moodreg, BMRQ_emotion,
                          BMRQ_socialreward, BMRQ_musicseek, BMRQ_sensorimotor,QUIC_child_total, 
                          CHAOS_child_total, deprivation_child_total, threat_childhood_total)
output = corr.test(variables)

corr_matrix = output$r
p_matrix <- output$p

colnames(corr_matrix) <- c('1', '2', '3', '4', '5', '6', '7', "8", "9", "10", "11", "12",'13', '14', '15', '16', '17', '18',
                         '19',"20", "21", "22")
rownames(corr_matrix) <- c('1', '2', '3', '4', '5', '6', '7', "8", "9", "10", "11", "12",'13', '14', '15', '16', '17', '18',
                         '19',"20", "21", "22")


pdf(file = "../../figures/pleasing:displeasing_followup/EFA_corrplot_totalsnobold.pdf", width = 13, height=13)

## update to get corrplot2
corrplot(corr_matrix, p.mat = p_matrix, insig = "label_sig",
         sig.level = c(0.001,0.01,0.05), pch.cex=0.9, pch.col = "black", tl.col = "black", type="lower",tl.srt =45, 
         col=colorRampPalette(c("darkblue","white","red"))(100),cl.lim=c(0,1), diag = FALSE,tl.cex = 1, cl.cex = 1)
dev.off()

#new construct validity
variables = dplyr::select(all, 
                          MH_total,
                          MH_caregiverbeliefs, MH_caregiversinging,
                          MH_childhoodattitude, MH_musicmaking, 
                          MH_socialcontexts, SRQ_adol_total, SRQ_current_total,
                          MET_melodicscore, MET_rhythmicscore,
                          What.is.your.age.)

output = corr.test(variables)
corr_matrix <- output$r
p_matrix <- output$p

colnames(corr_matrix) <- c('1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11')
rownames(corr_matrix) <- c('1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11')
corrplot(corr_matrix, p.mat = p_matrix, insig = "label_sig",
         sig.level = c(0.001,0.01,0.05), pch.cex=0.9, pch.col = "black", tl.col = "black", type="lower",tl.srt =45, 
         col=colorRampPalette(c("darkblue","white","red"))(100),cl.lim=c(0,1))


#utility of scale
variables = dplyr::select(all, 
                          MH_total,
                          MH_caregiverbeliefs, MH_caregiversinging,
                          MH_childhoodattitude, MH_musicmaking, 
                          MH_socialcontexts,
                          MET_melodicscore, MET_rhythmicscore,
                          HUMS_adult_healthy, HUMS_adult_unhealthy, 
                          STAI_total, CD.RISC_total)

output = corr.test(variables)
corr_matrix <- output$r
p_matrix <- output$p

colnames(corr_matrix) <- c('1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11', '12')
rownames(corr_matrix) <- c('1', '2', '3', '4', '5', '6', '7', '8', '9', '10', '11', '12')
corrplot(corr_matrix, p.mat = p_matrix, insig = "label_sig",
         sig.level = c(0.001,0.01,0.05), pch.cex=0.9, pch.col = "black", tl.col = "black", type="lower",tl.srt =45, 
         col=colorRampPalette(c("darkblue","white","red"))(100),cl.lim=c(0,1))

cor.test()

```

#updated figures for manuscript 
```{r}
#Replications GoldMSI
variables = dplyr::select(second, 
                          MH_total, MH_caregiverbeliefs, MH_caregiversinging, MH_childhoodattitude, MH_socialcontexts,
                          MH_musicmaking, Gold_activeengagement, Gold_perceptualabilities, Gold_musicaltraining,
                          Gold_singingabilities, Gold_emotion, Gold_musicalsophistication)
output = corr.test(variables)

corr_matrix = output$r
p_matrix <- output$p

colnames(corr_matrix) <- c('M@HR General', 'M@HR Caregiver Beliefs', 'M@HR Caregiver Singing', 'M@HR Attitude', 'M@HR Social Contexts', 'M@HR Child Engagement', 'Active Engagment', "Perceptual Abilities", "Musical Training", "Singing Abilities", "Emotion", "Musical Sophistication")
rownames(corr_matrix) <- c('M@HR General', 'M@HR Caregiver Beliefs', 'M@HR Caregiver Singing', 'M@HR Attitude', 'M@HR Social Contexts', 'M@HR Child Engagement', 'Active Engagment', "Perceptual Abilities", "Musical Training", "Singing Abilities", "Emotion", "Musical Sophistication")

pdf(file = "../../figures/m@h/Gold_Study2.pdf")

## update to get corrplot2
corrplot(corr_matrix, p.mat = p_matrix, insig = "label_sig",
         sig.level = c(0.001,0.01,0.05), pch.cex=0.9, pch.col = "black", tl.col = "black", type="lower",tl.srt =45, 
         col=colorRampPalette(c("darkblue","white","red"))(100),cl.lim=c(0,1), diag = FALSE,tl.cex = 1, cl.cex = 1)

dev.off()

#Replications BMRQ 
variables = dplyr::select(second, 
                          MH_total, MH_caregiverbeliefs, MH_caregiversinging, MH_childhoodattitude, MH_socialcontexts,
                          MH_musicmaking, BMRQ_total, BMRQ_moodreg, BMRQ_emotion,
                          BMRQ_socialreward, BMRQ_musicseek, BMRQ_sensorimotor)
output = corr.test(variables)

corr_matrix = output$r
p_matrix <- output$p

colnames(corr_matrix) <- c('M@HR General', 'M@HR Caregiver Beliefs', 'M@HR Caregiver Singing', 'M@HR Attitude', 'M@HR Social Contexts', 'M@HR Child Engagement', 'BMRQ Total', "Mood Regulation", "Emotion", "Social Reward", "Music Seeking", "Sensorimotor")
rownames(corr_matrix) <- c('M@HR General', 'M@HR Caregiver Beliefs', 'M@HR Caregiver Singing', 'M@HR Attitude', 'M@HR Social Contexts', 'M@HR Child Engagement', 'BMRQ Total', "Mood Regulation", "Emotion", "Social Reward", "Music Seeking", "Sensorimotor")

pdf(file = "../../figures/m@h/BMRQ_Study2.pdf")

## update to get corrplot2
corrplot(corr_matrix, p.mat = p_matrix, insig = "label_sig",
         sig.level = c(0.001,0.01,0.05), pch.cex=0.9, pch.col = "black", tl.col = "black", type="lower",tl.srt =45, 
         col=colorRampPalette(c("darkblue","white","red"))(100),cl.lim=c(0,1), diag = FALSE,tl.cex = 1, cl.cex = 1)

dev.off()

#Adversity 
variables = dplyr::select(second, 
                          MH_total, MH_caregiverbeliefs, MH_caregiversinging, MH_childhoodattitude, MH_socialcontexts,
                          MH_musicmaking, QUIC_child_total, deprivation_child_total, threat_childhood_total, CHAOS_child_total)
output = corr.test(variables)

corr_matrix = output$r
p_matrix <- output$p

colnames(corr_matrix) <- c('M@HR General', 'M@HR Caregiver Beliefs', 'M@HR Caregiver Singing', 'M@HR Attitude', 'M@HR Social Contexts', 'M@HR Child Engagement', 'QUIC', "Deprivation", "Threat", "CHAOS")
rownames(corr_matrix) <- c('M@HR General', 'M@HR Caregiver Beliefs', 'M@HR Caregiver Singing', 'M@HR Attitude', 'M@HR Social Contexts', 'M@HR Child Engagement', 'QUIC', "Deprivation", "Threat", "CHAOS")

pdf(file = "../../figures/m@h/Adversity_Study2.pdf")

## update to get corrplot2
corrplot(corr_matrix, p.mat = p_matrix, insig = "label_sig",
         sig.level = c(0.001,0.01,0.05), pch.cex=0.9, pch.col = "black", tl.col = "black", type="lower",tl.srt =45, 
         col=colorRampPalette(c("darkblue","white","red"))(100),cl.lim=c(0,1), diag = FALSE,tl.cex = 1, cl.cex = 1)

dev.off()

#SRQ
variables = dplyr::select(all, 
                          MH_total, MH_caregiverbeliefs, MH_caregiversinging, MH_childhoodattitude, MH_socialcontexts,
                          MH_musicmaking, SRQ_current_total)
output = corr.test(variables)

corr_matrix = output$r
p_matrix <- output$p

colnames(corr_matrix) <- c('M@HR General', 'M@HR Caregiver Beliefs', 'M@HR Caregiver Singing', 'M@HR Attitude', 'M@HR Social Contexts', 'M@HR Child Engagement', 'SRQ')
rownames(corr_matrix) <- c('M@HR General', 'M@HR Caregiver Beliefs', 'M@HR Caregiver Singing', 'M@HR Attitude', 'M@HR Social Contexts', 'M@HR Child Engagement', 'SRQ')

pdf(file = "../../figures/m@h/SRQ_Study2.pdf")

## update to get corrplot2
corrplot(corr_matrix, p.mat = p_matrix, insig = "label_sig",
         sig.level = c(0.001,0.01,0.05), pch.cex=0.9, pch.col = "black", tl.col = "black", type="lower",tl.srt =45, 
         col=colorRampPalette(c("darkblue","white","red"))(100),cl.lim=c(0,1), diag = FALSE,tl.cex = 1, cl.cex = 1)

dev.off()

#MET
variables = dplyr::select(all, 
                          MH_total, MH_caregiverbeliefs, MH_caregiversinging, MH_childhoodattitude, MH_socialcontexts,
                          MH_musicmaking, MET_melodicscore, MET_rhythmicscore)
output = corr.test(variables)

corr_matrix = output$r
p_matrix <- output$p

colnames(corr_matrix) <- c('M@HR General', 'M@HR Caregiver Beliefs', 'M@HR Caregiver Singing', 'M@HR Attitude', 'M@HR Social Contexts', 'M@HR Child Engagement', 'MET Melodic', 'MET Rhythmic')
rownames(corr_matrix) <- c('M@HR General', 'M@HR Caregiver Beliefs', 'M@HR Caregiver Singing', 'M@HR Attitude', 'M@HR Social Contexts', 'M@HR Child Engagement', 'MET Melodic', 'MET Rhythmic')

pdf(file = "../../figures/m@h/MET_Study2.pdf")

## update to get corrplot2
corrplot(corr_matrix, p.mat = p_matrix, insig = "label_sig",
         sig.level = c(0.001,0.01,0.05), pch.cex=0.9, pch.col = "black", tl.col = "black", type="lower",tl.srt =45, 
         col=colorRampPalette(c("darkblue","white","red"))(100),cl.lim=c(0,1), diag = FALSE,tl.cex = 1, cl.cex = 1)

dev.off()

#HUMS
variables = dplyr::select(all, 
                          MH_total, MH_caregiverbeliefs, MH_caregiversinging, MH_childhoodattitude, MH_socialcontexts,
                          MH_musicmaking, HUMS_adult_healthy, HUMS_adult_unhealthy)
output = corr.test(variables)

corr_matrix = output$r
p_matrix <- output$p

colnames(corr_matrix) <- c('M@HR General', 'M@HR Caregiver Beliefs', 'M@HR Caregiver Singing', 'M@HR Attitude', 'M@HR Social Contexts', 'M@HR Child Engagement', 'HUMS Healthy', 'HUMS Unhealthy')
rownames(corr_matrix) <- c('M@HR General', 'M@HR Caregiver Beliefs', 'M@HR Caregiver Singing', 'M@HR Attitude', 'M@HR Social Contexts', 'M@HR Child Engagement', 'HUMS Healthy', 'HUMS Unhealthy')

pdf(file = "../../figures/m@h/HUMS_Study2.pdf")

## update to get corrplot2
corrplot(corr_matrix, p.mat = p_matrix, insig = "label_sig",
         sig.level = c(0.001,0.01,0.05), pch.cex=0.9, pch.col = "black", tl.col = "black", type="lower",tl.srt =45, 
         col=colorRampPalette(c("darkblue","white","red"))(100),cl.lim=c(0,1), diag = FALSE,tl.cex = 1, cl.cex = 1)

dev.off()

#Mental Health
variables = dplyr::select(all, 
                          MH_total, MH_caregiverbeliefs, MH_caregiversinging, MH_childhoodattitude, MH_socialcontexts,
                          MH_musicmaking, STAI_total, CD.RISC_total)
output = corr.test(variables)

corr_matrix = output$r
p_matrix <- output$p

colnames(corr_matrix) <- c('M@HR General', 'M@HR Caregiver Beliefs', 'M@HR Caregiver Singing', 'M@HR Attitude', 'M@HR Social Contexts', 'M@HR Child Engagement', 'STAI', 'CD.RISC')
rownames(corr_matrix) <- c('M@HR General', 'M@HR Caregiver Beliefs', 'M@HR Caregiver Singing', 'M@HR Attitude', 'M@HR Social Contexts', 'M@HR Child Engagement', 'STAI', 'CD.RISC')

pdf(file = "../../figures/m@h/mentalhealth_Study2.pdf")

## update to get corrplot2
corrplot(corr_matrix, p.mat = p_matrix, insig = "label_sig",
         sig.level = c(0.001,0.01,0.05), pch.cex=0.9, pch.col = "black", tl.col = "black", type="lower",tl.srt =45, 
         col=colorRampPalette(c("darkblue","white","red"))(100),cl.lim=c(0,1), diag = FALSE,tl.cex = 1, cl.cex = 1)

dev.off()

#Utility of Scale Scatterplots
MET_melodic_general <-
ggplot(all, aes(x=MH_total, y=MET_melodicscore)) + 
  theme_classic() +
  geom_point() + 
  geom_smooth(method=lm) + 
  xlab("M@HR General") +
  ylab("MET Melodic") +
    scale_x_continuous(breaks = scales::pretty_breaks(n = 5)) +
  scale_y_continuous(breaks = scales::pretty_breaks(n = 5)) +
  theme(axis.title=element_text(size=20,face="bold"), axis.text=element_text(size=15))
print(MET_melodic_general)
ggsave("Figure3_A.png", width = 6.5, height = 4)

MET_melodic_cbeliefs <-
ggplot(all, aes(x=MH_caregiverbeliefs, y=MET_melodicscore)) + 
  theme_classic() +
  geom_point(position="jitter") + 
  geom_smooth(method=lm) + 
  xlab("M@HR Caregiver Beliefs") +
  ylab("MET Melodic") +
  scale_x_continuous(breaks = scales::pretty_breaks(n = 5)) +
  scale_y_continuous(breaks = scales::pretty_breaks(n = 5)) +
  theme(axis.title=element_text(size=20,face="bold"), axis.text=element_text(size=15))
print(MET_melodic_cbeliefs)
ggsave("Figure3_B.png", width = 6.5, height = 4)

MET_melodic_childengagement <-
ggplot(all, aes(x=MH_musicmaking, y=MET_melodicscore)) + 
  theme_classic() +
  geom_point(position="jitter") + 
  geom_smooth(method=lm) + 
  xlab("M@HR Child Engagement") +
  ylab("MET Melodic") +
  scale_x_continuous(breaks = scales::pretty_breaks(n = 5)) +
  scale_y_continuous(breaks = scales::pretty_breaks(n = 5)) +
  theme(axis.title=element_text(size=20,face="bold"), axis.text=element_text(size=15))
print(MET_melodic_childengagement)
ggsave("Figure3_C.png", width = 6.5, height = 4)

MH_total <- all$MH_total
MET_melodicscore <- all$MET_melodicscore
corr.test(MH_total, MET_melodicscore)

HUMS_healthy <-
ggplot(all, aes(x=MH_total, y=HUMS_adult_healthy)) + 
  theme_classic() +
  geom_point(position="jitter") + 
  geom_smooth(method=lm) + 
  xlab("M@HR General") +
  ylab("HUMS Healthy") +
  scale_x_continuous(breaks = scales::pretty_breaks(n = 5)) +
  scale_y_continuous(breaks = scales::pretty_breaks(n = 5)) +
  theme(axis.title=element_text(size=20,face="bold"), axis.text=element_text(size=15))
print(HUMS_healthy)
ggsave("Figure4_A.png", width = 6.5, height = 4)

HUMS_unhealthy <-
ggplot(all, aes(x=MH_childhoodattitude, y=HUMS_adult_unhealthy)) + 
  theme_classic() +
  geom_point(position="jitter") + 
  geom_smooth(method=lm) + 
  xlab("M@HR Attitude") +
  ylab("HUMS Unhealthy") +
  scale_x_continuous(breaks = scales::pretty_breaks(n = 5)) +
  scale_y_continuous(breaks = scales::pretty_breaks(n = 5)) +
  theme(axis.title=element_text(size=20,face="bold"), axis.text=element_text(size=15))
print(HUMS_unhealthy)
ggsave("Figure4_B.png", width = 6.5, height = 4)

STAI <-
ggplot(all, aes(x=MH_total, y=STAI_total)) + 
  theme_classic() +
  geom_point() + 
  geom_smooth(method=lm) + 
  xlab("M@HR General") +
  ylab("STAI Trait Anxiety") +
  scale_x_continuous(breaks = scales::pretty_breaks(n = 5)) +
  scale_y_continuous(breaks = scales::pretty_breaks(n = 5)) +
  theme(axis.title=element_text(size=20,face="bold"), axis.text=element_text(size=15))
print(STAI)
ggsave("Figure5_A.png", width = 6.5, height = 4)

STAI <-
ggplot(all, aes(x=MH_caregiversinging, y=STAI_total)) + 
  theme_classic() +
  geom_point(position="jitter") + 
  geom_smooth(method=lm) + 
  xlab("M@HR Caregiver Singing") +
  ylab("STAI Trait Anxiety") +
  scale_x_continuous(breaks = scales::pretty_breaks(n = 5)) +
  scale_y_continuous(breaks = scales::pretty_breaks(n = 5)) +
  theme(axis.title=element_text(size=20,face="bold"), axis.text=element_text(size=15))
print(STAI)
ggsave("Figure5_B.png", width = 6.5, height = 4)

STAI <-
ggplot(all, aes(x=MH_childhoodattitude, y=STAI_total)) + 
  theme_classic() +
  geom_point(position="jitter") + 
  geom_smooth(method=lm) + 
  xlab("M@HR Attitude") +
  ylab("STAI Trait Anxiety") +
  scale_x_continuous(breaks = scales::pretty_breaks(n = 5)) +
  scale_y_continuous(breaks = scales::pretty_breaks(n = 5)) +
  theme(axis.title=element_text(size=20,face="bold"), axis.text=element_text(size=15))
print(STAI)
ggsave("Figure5_C.png", width = 6.5, height = 4)

STAI_total <- all$STAI_total
corr.test(MH_total, STAI_total)

CD.RISC <-
ggplot(all, aes(x=MH_total, y=CD.RISC_total)) + 
  theme_classic() +
  geom_point() + 
  geom_smooth(method=lm) + 
  xlab("M@HR General") +
  ylab("CD.RISC Resilience") +
  scale_x_continuous(breaks = scales::pretty_breaks(n = 5)) +
  scale_y_continuous(breaks = scales::pretty_breaks(n = 5)) +
  theme(axis.title=element_text(size=20,face="bold"), axis.text=element_text(size=15))
print(CD.RISC)
ggsave("Figure6_A.png", width = 6.5, height = 4)

CD.RISC <-
ggplot(all, aes(x=MH_caregiverbeliefs, y=CD.RISC_total)) + 
  theme_classic() +
  geom_point(position="jitter") + 
  geom_smooth(method=lm) + 
  xlab("M@HR Caregiver Beliefs") +
  ylab("CD.RISC Resilience") +
  scale_x_continuous(breaks = scales::pretty_breaks(n = 5)) +
  scale_y_continuous(breaks = scales::pretty_breaks(n = 5)) +
  theme(axis.title=element_text(size=20,face="bold"), axis.text=element_text(size=15))
print(CD.RISC)
ggsave("Figure6_B.png", width = 6.5, height = 4)

CD.RISC <-
ggplot(all, aes(x=MH_caregiversinging, y=CD.RISC_total)) + 
  theme_classic() +
  geom_point(position="jitter") + 
  geom_smooth(method=lm) + 
  xlab("M@HR Caregiver Singing") +
  ylab("CD.RISC Resilience") +
  scale_x_continuous(breaks = scales::pretty_breaks(n = 5)) +
  scale_y_continuous(breaks = scales::pretty_breaks(n = 5)) +
  theme(axis.title=element_text(size=20,face="bold"), axis.text=element_text(size=15))
print(CD.RISC)
ggsave("Figure6_C.png", width = 6.5, height = 4)

CD.RISC <-
ggplot(all, aes(x=MH_socialcontexts, y=CD.RISC_total)) + 
  theme_classic() +
  geom_point(position="jitter") + 
  geom_smooth(method=lm) + 
  xlab("M@HR Social Listening Contexts") +
  ylab("CD.RISC Resilience") +
  scale_x_continuous(breaks = scales::pretty_breaks(n = 5)) +
  scale_y_continuous(breaks = scales::pretty_breaks(n = 5)) +
  theme(axis.title=element_text(size=20,face="bold"), axis.text=element_text(size=15))
print(CD.RISC)
ggsave("Figure6_D.png", width = 6.5, height = 4)

CD.RISC <-
ggplot(all, aes(x=MH_childhoodattitude, y=CD.RISC_total)) + 
  theme_classic() +
  geom_point(position="jitter") + 
  geom_smooth(method=lm) + 
  xlab("M@HR Attitude") +
  ylab("CD.RISC Resilience") +
  scale_x_continuous(breaks = scales::pretty_breaks(n = 5)) +
  scale_y_continuous(breaks = scales::pretty_breaks(n = 5)) +
  theme(axis.title=element_text(size=20,face="bold"), axis.text=element_text(size=15))
print(CD.RISC)
ggsave("Figure6_E.png", width = 6.5, height = 4)

SRQ<-
ggplot(all, aes(x=MH_total, y=SRQ_current_total)) + 
  theme_classic() +
  geom_point() + 
  geom_smooth(method=lm) + 
  xlab("M@HR General") +
  ylab("SRQ") +
  theme(axis.title=element_text(size=20,face="bold"), axis.text=element_text(size=15))
print(SRQ)
ggsave("SRQ.png", width = 6.5, height = 4)

SRQ<-
ggplot(all, aes(x=MH_socialcontexts, y=SRQ_current_total)) + 
  theme_classic() +
  geom_point(position="jitter") + 
  geom_smooth(method=lm) + 
  xlab("M@HR Social Listening Contexts") +
  ylab("SRQ Total") +
  scale_x_continuous(breaks = scales::pretty_breaks(n = 5)) +
  scale_y_continuous(breaks = scales::pretty_breaks(n = 5)) +
  theme(axis.title=element_text(size=20,face="bold"), axis.text=element_text(size=15))
print(SRQ)
ggsave("Figure2_B.png", width = 6.5, height = 4)

SRQ_current_total <- all$SRQ_current_total
corr.test(MH_total, SRQ_current_total)

BMRQ_socialreward <-
ggplot(all, aes(x=MH_socialcontexts, y=BMRQ_socialreward)) + 
  theme_classic() +
  geom_point(position="jitter") + 
  geom_smooth(method=lm) + 
  xlab("M@HR Social Listening Contexts") +
  ylab("eBMRQ Music Social Reward") +
  scale_x_continuous(breaks = scales::pretty_breaks(n = 5)) +
  scale_y_continuous(breaks = scales::pretty_breaks(n = 5)) +
  theme(axis.title=element_text(size=20,face="bold"), axis.text=element_text(size=15))
print(BMRQ_socialreward)
ggsave("Figure2_A.png", width = 6.5, height = 4.4)


```


#Utility of scale models
```{r}
m1 <- lm(scale(HUMS_adult_healthy) ~ scale(MH_caregiverbeliefs) + scale(MH_musicmaking) + scale(MH_caregiversinging) + scale(MH_socialcontexts) + scale(MH_childhoodattitude), data = all)
summary(m1)

HUMS_adult_healthy <- 
plot_model(m1, show.values = TRUE,vline.color = "lightgrey", colors="red") + 
  ylim(-.5, .5) +
  ggtitle("HUMS Healthy") +
  ylab("Standardized Beta Coefficient Estimates") +
  xlab("M@HR") +
  scale_x_discrete(labels=c("Attitude","Social Listening Contexts", "Caregiver Singing", "Child Engagement", "Caregiver Beliefs")) + 
  theme_light() +
  geom_hline(yintercept=0,color='black',linetype='dashed') +
  theme(axis.title=element_text(size=15, face="bold"), axis.text=element_text(size=15, face="bold"), plot.title=element_text(size=15, face="bold"), axis.text.x=element_text(size=10))

print(HUMS_adult_healthy)
ggsave("Figure4_C.png", width = 6.85, height = 4)

m1 <- lm(PASS_total ~ MH_caregiverbeliefs + MH_caregiversinging + MH_childhoodattitude + MH_musicmaking + MH_socialcontexts, data = all)
summary(m1)

m1 <- lm(STAI_total ~ MH_caregiverbeliefs + MH_caregiversinging + MH_childhoodattitude + MH_musicmaking + MH_socialcontexts, data = all)
summary(m1)

m1 <- lm(scale(MH_socialcontexts) ~ scale(BMRQ_moodreg) + scale(BMRQ_emotion) + scale(BMRQ_socialreward) + scale(BMRQ_absorption) + scale(BMRQ_musicseek) + scale(BMRQ_sensorimotor), data=all)
summary(m1)

BMRQ_socialreward <- 
plot_model(m1, show.values = TRUE,vline.color = "lightgrey", colors=c("blue","red")) + 
  ylim(-.5, .5) +
  ggtitle("M@HR Social Listening Contexts") +
  ylab("Standardized Beta Coefficient Estimates") +
  xlab("eBMRQ") +
  scale_x_discrete(labels=c("Sensorimotor","Music Seeking", "Absorption", "Social Reward", "Emotion", "Mood Regulation")) + 
  theme_light() +
  geom_hline(yintercept=0,color='black',linetype='dashed') +
  theme(axis.title=element_text(size=15, face="bold"), axis.text=element_text(size=15, face="bold"), plot.title=element_text(size=15, face="bold"), axis.text.x=element_text(size=10))

print(BMRQ_socialreward)
ggsave("Figure2_C.png", width = 6.85, height = 4)

m1 <- lm(scale(SRQ_current_total) ~ scale(MH_caregiverbeliefs) + scale(MH_musicmaking) + scale(MH_caregiversinging) + scale(MH_socialcontexts) + scale(MH_childhoodattitude), data = all)
summary(m1)

SRQ <- 
plot_model(m1, show.values = TRUE,vline.color = "lightgrey", colors = 'red') + 
  ylim(-.5, .5) +
  ggtitle("SRQ") +
  ylab("Standardized Beta Coefficient Estimates") +
  xlab ("M@HR") +
  scale_x_discrete(labels=c("Attitude","Social Listening Contexts", "Caregiver Singing", "Child Engagement", "Caregiver Beliefs")) + 
  theme_light() +
  geom_hline(yintercept=0,color='black',linetype='dashed') +
  theme(axis.title=element_text(size=15, face="bold"), axis.text=element_text(size=15, face="bold"), plot.title=element_text(size=15, face="bold"), axis.text.x=element_text(size=10))
print(SRQ)
ggsave("Figure2_D.png", width = 6.85, height = 4)

##scatter plots##
ggplot(data=all, aes(x=MH_total, y=CD.RISC_total))+
  geom_point()+
  stat_smooth(method="lm") +
  theme_classic()

#checking if MET/STAI/cd.risc correlations persist w. musical training as a covariate
m1 <- lm(HUMS_adult_healthy ~ MH_total + Gold_musicaltraining, data = all)
summary(m1)
all$HUMS_adult_healthy
```

```{r}
mh <- all$MH_childhoodattitude
x <- all$BMRQ_musicseek

cor.test(mh, x)
```

#tables
```{r}
#Test-Retest
testretest_table <- testretest_table %>%
    dplyr::rename_with(
      ~ case_when(
        . == "Music.Home...Retrospective.Subscales" ~ "Music@Home - Retrospective Subscales ",
        . == "Test.Retest..ICC..Over.4.Months" ~ "Test-Retest (ICC) Over 4 Months",
        TRUE ~ .
      )
    )

testretest_table_apa <-
  nice_table(
    testretest_table[1:8, ],
    title = c("Table 3", "Test-Retest Reliability"),
    note = c("Music@Home - Retrospective test-retest reliability measured via intra-class correlation (ICC) values between Study 1 and Study 2 (4 months delay) for 46 participants. The following thresholds were used: fair—0.40 ≤ values ≤ 0.59; good— 0.60 ≤ values ≤ 0.74; and excellent— values ≥ 0.75 for test-retest reliability.")
  )
print(testretest_table_apa)
flextable::save_as_docx(testretest_table_apa, path = '../../figures/m@h/testretest_table.docx')
  
#Gold-MSI
goldmsi_table <- goldmsi_table %>%
    dplyr::rename_with(
      ~ case_when(
        . == "Music.Home...Retrospective.Subscales" ~ " ",
        . == "Active.Engagement" ~ "Active Engagement",
        . == "Perceptual.Abilities" ~ "Perceptual Abilities",
        . == "Musical.Training" ~ "Musical Training",
        . == "Singing.Abilities" ~ "Singing Abilities",
        . == "Emotion" ~ "Emotion",
        . == "Musical.Sophistication" ~ "Musical Sophistication",
        TRUE ~ .
      )
    )

goldmsi_table_apa <-
  nice_table(
    goldmsi_table[1:7, ],
    title = c("Table 4", "Gold-MSI Subscale Associations"),
    note = c("Bivariate correlations relating Goldsmith Musical Sophistication Index (Gold-MSI) subscales (top row) and Music@Home - Retrospective subscales (left column) in the Study 2 sample. * p < .05, ** p <.01, *** p <.001.")
  )
print(goldmsi_table_apa)
flextable::save_as_docx(goldmsi_table_apa, path = '../../figures/m@h/goldmsi_table.docx')

#eBMRQ
ebmrq_table <- ebmrq_table %>%
    dplyr::rename_with(
      ~ case_when(
        . == "Music.Home...Retrospective.Subscales" ~ " ",
        . == "Mood.Regulation" ~ "Mood Regulation",
        . == "Emotion" ~ "Emotion",
        . == "Social.Reward" ~ "Social Reward",
        . == "Absorption" ~ "Absorption",
        . == "Music.Seeking" ~ "Music Seeking",
        . == "Sensorimotor" ~ "Sensorimotor",
        . == "Total" ~ "Total",
        TRUE ~ .
      )
    )

ebmrq_table_apa <-
  nice_table(
    ebmrq_table[1:8, ],
    title = c("Table 5", "eBMRQ Subscale Associations"),
    note = c("Bivariate correlations relating Extended Barcelona Music Reward Questionnaire subscales (eBMRQ) subscales (top row) and Music@Home - Retrospective subscales (left column) in the Study 2 sample. * p < .05, ** p <.01, *** p <.001.")
  )
print(ebmrq_table_apa)
flextable::save_as_docx(ebmrq_table_apa, path = '../../figures/m@h/ebmrq_table.docx')

#SRQ
SRQ_table <- SRQ_table %>%
    dplyr::rename_with(
      ~ case_when(
        . == "Music.Home...Retrospective.Subscales" ~ "Music@Home - Retrospective Subscales ",
        . == "SRQ" ~ "SRQ",
        TRUE ~ .
      )
    )

SRQ_table_apa <-
  nice_table(
    SRQ_table[1:8, ],
    title = c("Table 6", "SRQ Associations"),
    note = c("Bivariate correlations relating the Social Reward Questionnaire (SRQ) subscales and Music@Home - Retrospective subscales in the Study 2 sample. * p < .05, ** p <.01, *** p <.001.")
  )
print(SRQ_table_apa)
flextable::save_as_docx(SRQ_table_apa, path = '../../figures/m@h/SRQ_table.docx')

#study 1 demo
demo_table1 <- demo_table1 %>%
    dplyr::rename_with(
      ~ case_when(
        . == "X" ~ "Study 1 Sample (n = 290)",
        TRUE ~ .
      )
    )

demo1table <- 
nice_table(
    demo_table1[1:19, ],
   title = c("Table 1", "Study 1 Sample Demographics"),
    note = c("Demographic information for the Study 1 sample.")
  )

flextable::save_as_docx(demo1table, path = '../../figures/m@h/demo1table.docx')

#study 2 demo
demo_table <- demo_table %>%
    dplyr::rename_with(
      ~ case_when(
        . == "Entire.Study.2.Sample.n...288" ~ "Entire Study 2 Sample (n = 288)",
        . == "Test.Retest.Sample.n...46" ~ "Test Retest Sample (n = 46)",
        TRUE ~ .
      )
    )

#demo2table <- 
nice_table(
    demo_table[1:19, ],
   title = c("Table 5", "Study 2 Sample Demographics"),
    note = c("Demographic information for the entire Study 2 sample, as well as only those included in the test-retest analysis.")
  )

flextable::save_as_docx(demo2table, path = '../../figures/m@h/demo2table.docx')
```
